refactor: move gallery component into its own directory
